fix DI issue at expense of breaking tests again [#135473733]

+++ app/org/birdfeed/chirp/initializers/Permissions.scala
@@ -1,9 +1,13 @@
 package org.birdfeed.chirp.initializers
 
+import akka.actor.ActorSystem
 import com.google.inject.{Inject, Singleton}
 import org.birdfeed.chirp.database.models.{Permission, Role}
-import play.api.Application
+import play.api.{Application, Configuration, Environment, Mode}
 import play.api.db.Database
+import com.github.aselab.activerecord._
+import com.github.aselab.activerecord.dsl._
+import org.birdfeed.chirp.database.{SchemaTables} /* Do not remove, it scopes an implicit value */
 
 /**
   * Map permissions from models.
@@ -12,6 +16,11 @@ import play.api.db.Database
   * hopefully Scala-ActiveRecord keeps a list of some sort
   *
   * http://stackoverflow.com/questions/19118283/how-can-i-get-all-object-vals-and-subobject-vals-using-reflection-in-scala
+  *
+  * Q: Why do we DI so many things?
+  * A: The anatomy of a play application is not monolithic. We
+  * have to wait for larger components to do some database operations.
+  *
   * @param app Initialized Play Application
   *            (this is important as it ensures everything
   *            to do with the DB connection pool is
@@ -20,41 +29,64 @@ import play.api.db.Database
   *            is ready)
   */
 @Singleton
-class Permissions @Inject()(val app: Application, val db: Database) {
+class Permissions @Inject()(
+  val app: Application,
+  val db: Database,
+  actorSystem: ActorSystem,
+  conf: Configuration,
+  env: Environment,
+  permission: Permission,
+  role: Role
+) {
+
   Permission.deleteAll
   Role.deleteAll
-
 
   // On self hatred and disgust
   db.withConnection { conn =>
     // Scala-ActiveRecord seems to not create a `serial`
     // so we don't get restart identity
     // TODO: Pull request!
-    conn.createStatement.execute(
-      conn.nativeSQL("ALTER SEQUENCE permissions_id_seq RESTART WITH 1;") +
-        conn.nativeSQL("ALTER SEQUENCE roles_id_seq RESTART WITH 1;") +
-        conn.nativeSQL("TRUNCATE TABLE permissions_roles;")
-    )
+
+    if (env.mode != Mode.Test) {
+      conn.createStatement.execute(
+        conn.nativeSQL("ALTER SEQUENCE permissions_id_seq RESTART WITH 1;") +
+          conn.nativeSQL("ALTER SEQUENCE roles_id_seq RESTART WITH 1;") +
+          conn.nativeSQL("TRUNCATE TABLE permissions_roles;")
+      )
+    }
+
   }
+
 
   val permissionable = Seq(
     "user", "accessToken", "apiKey", "sample", "experiment", "score"
   ).map { permission =>
     Seq(
-      Permission(permission).create,
-      Permission(s"$permission.write").create
+      Permission(permission),
+      Permission(s"$permission.write")
     )
   }
 
-  val administrator = Role("Administrator")
-  val researcher = Role("Researcher")
-  val participant = Role("Participant")
+//  val administrator = Role("Administrator")
+//  val researcher = Role("Researcher")
+//  val participant = Role("Participant")
 
-  administrator.create.permissions ++= permissionable.flatten
-  researcher.create.permissions ++= permissionable.slice(3, 5).flatten
-  participant.create.permissions ++= permissionable.last
+  var roles = Seq(
+    Role("Administrator").create,
+    Role("Researcher").create,
+    Role("Participant").create
+  )
 
-  administrator.save
-  researcher.save
-  participant.save
+  Role.inTransaction {
+    roles = roles.map(SchemaTables.roles.insert)
+  }
+
+  roles(0).permissions ++= permissionable.flatten
+  roles(1).permissions ++= permissionable.slice(3, 5).flatten
+  roles(2).permissions ++= permissionable.last
+
+  roles(0).save
+  roles(1).save
+  roles(2).save
 }
